requirements.txt
audioread==2.1.9
einops==0.3.2
fire==0.4.0
h5py==3.6.0
ignite==1.1.0
librosa==0.8.1
loguru==0.6.0
numpy==1.21.5
pandas==1.4.0
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.7.3
SoundFile==0.10.3.post1
torch==1.10.2
torch_audiomentations==0.9.1
torchaudio==0.10.1+4b64f80
tqdm==4.62.3
